requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosmtplib==3.0.2
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.1
attrs==24.2.0
Babel==2.14.0
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
crypto==1.4.1
cryptography==42.0.0
dataclasses-json==0.5.8
dateutils==0.6.12
dnspython==2.4.2
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.16.7
email-validator==2.1.0.post1
fastapi==0.104.1
fastapi-pagination==0.12.14
frozenlist==1.4.1
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
idna==3.4
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.4
marshmallow==3.21.3
marshmallow-enum==1.5.1
multidict==6.0.5
mypy-extensions==1.0.0
Naked==0.1.32
numpy==1.26.1
packaging==24.1
pandas==2.1.2
passlib==1.7.4
pdfrw==0.4
phonepe-sdk==1.1.0
pillow==10.2.0
ply==3.11
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
razorpay==1.4.2
redis==5.0.1
reportlab==4.0.9
requests==2.31.0
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
tqdm==4.66.1
twilio==9.3.0
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.12
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.11
